+++ packages/charts/react-charting/src/components/ScatterChart/ScatterChart.base.tsx
@@ -560,13 +560,7 @@ export const ScatterChartBase: React.FunctionComponent<IScatterChartProps> = Rea
               xAxisScale: _xAxisScale.current,
               yAxisScale: _yAxisScale.current,
               className: classNames.markerLabel || '',
-              maybeLineOptions: (_points.current?.[i] as Partial<ILineChartPoints>)?.lineOptions
-                ? {
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    originXOffset: ((_points.current?.[i] as Partial<ILineChartPoints>).lineOptions as any)
-                      ?.originXOffset,
-                  }
-                : undefined,
+              lineOptions: (_points.current?.[i] as Partial<ILineChartPoints>)?.lineOptions,
             }),
           );
         }