+++ packages/charts/react-charts/library/src/components/CommonComponents/CartesianChart.types.ts
@@ -1,9 +1,22 @@
 import * as React from 'react';
 import { LegendsProps } from '../Legends/index';
-import { AccessibilityProps, Chart, Margins, DataPoint, HorizontalBarChartWithAxisDataPoint } from '../../types/index';
-import { ChartTypes, XAxisTypes, YAxisType, IDomainNRange } from '../../utilities/index';
+import {
+  AccessibilityProps,
+  Chart,
+  Margins,
+  DataPoint,
+  HorizontalBarChartWithAxisDataPoint,
+  GroupedVerticalBarChartData,
+  HeatMapChartDataPoint,
+  LineChartPoints,
+  VerticalBarChartDataPoint,
+  VerticalStackedBarDataPoint,
+  ScatterChartDataPoint,
+} from '../../types/index';
 import { TimeLocaleDefinition } from 'd3-time-format';
 import { ChartPopoverProps } from './ChartPopover.types';
+import { ChartTypes, IAxisData, IDomainNRange, IYAxisParams, XAxisTypes, YAxisType } from '../../utilities/utilities';
+import { ScaleBand, ScaleLinear } from 'd3-scale';
 /**
  * Cartesian Chart style properties
  * {@docCategory CartesianChart}
@@ -589,17 +602,41 @@ export interface ModifiedCartesianChartProps extends CartesianChartProps {
   /**
    * Get the min and max values of the y-axis
    */
-  getMinMaxOfYAxis?: (
-    points: DataPoint[],
+  getMinMaxOfYAxis: (
+    points:
+      | LineChartPoints[]
+      | HorizontalBarChartWithAxisDataPoint[]
+      | VerticalBarChartDataPoint[]
+      | DataPoint[]
+      | ScatterChartDataPoint[],
     yAxisType: YAxisType | undefined,
     useSecondaryYScale?: boolean,
   ) => { startValue: number; endValue: number };
 
-  /**Add commentMore actions
+  /**
+   * Create the y-axis
+   */
+  createYAxis: (
+    yAxisParams: IYAxisParams,
+    isRtl: boolean,
+    axisData: IAxisData,
+    isIntegralDataset: boolean,
+    chartType: ChartTypes,
+    useSecondaryYScale?: boolean,
+    roundedTicks?: boolean,
+  ) => ScaleLinear<number, number, never>;
+
+  /**
    * Get the domain and range values
    */
-  getDomainNRangeValues?: (
-    points: HorizontalBarChartWithAxisDataPoint[],
+  getDomainNRangeValues: (
+    points:
+      | LineChartPoints[]
+      | VerticalBarChartDataPoint[]
+      | VerticalStackedBarDataPoint[]
+      | HorizontalBarChartWithAxisDataPoint[]
+      | GroupedVerticalBarChartData[]
+      | HeatMapChartDataPoint[],
     margins: Margins,
     width: number,
     chartType: ChartTypes,
@@ -609,4 +646,15 @@ export interface ModifiedCartesianChartProps extends CartesianChartProps {
     tickValues: Date[] | number[] | string[] | undefined,
     shiftX: number,
   ) => IDomainNRange;
+
+  /**
+   * Create the string y-axis
+   */
+  createStringYAxis: (
+    yAxisParams: IYAxisParams,
+    dataPoints: string[],
+    isRtl: boolean,
+    barWidth: number | undefined,
+    chartType?: ChartTypes,
+  ) => ScaleBand<string>;
 }